% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{data_graphic_draw_norm}
\alias{data_graphic_draw_norm}
\title{Draw Dataset Summary.}
\usage{
data_graphic_draw_norm(prot_dat, condition_num, repeat_num, state)
}
\arguments{
\item{prot_dat}{A DataFrame with only numeric value.}

\item{condition_num}{A integer of condition number.}

\item{repeat_num}{A integer of repeat number.}

\item{state}{A title}
}
\description{
Draw dataset summary with density distribution, pca and cv.
}
\details{
Input dataframe, then return a dataset summary with density distribution, pca and cv.
}
\examples{
prot_norm_example <- data_norm(prot_dat_example[1:15], 5, 3)
}
