% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{data_norm}
\alias{data_norm}
\title{Normalize MS TMT data.}
\usage{
data_norm(prot_dat, condition_num, repeat_num)
}
\arguments{
\item{prot_dat}{A DataFrame from read_maxquant_prot() with only numeric value.}

\item{condition_num}{A integer of condition number.}

\item{repeat_num}{A integer of repeat number.}
}
\value{
A DataFrame.

Plots.
}
\description{
Normalize MS TMT data which contain no zero rows.
}
\details{
Input dataframe from read_maxquant_prot() function, then return a data set after normalization with SL, TMM and IRS.
}
