% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{data_norm_sl}
\alias{data_norm_sl}
\title{Normalize MS TMT data with Sample Loading.}
\usage{
data_norm_sl(prot_dat)
}
\arguments{
\item{prot_dat}{A DataFrame from read_maxquant_prot() with only numeric value.}
}
\value{
A DataFrame.
}
\description{
Normalize MS TMT data with sample loading.
}
\details{
Input dataframe from read_maxquant_prot() function, then return a data set after normalization with SL.
}
