% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{data_norm_tmm}
\alias{data_norm_tmm}
\title{Normalize MS TMT data with TMM.}
\usage{
data_norm_tmm(prot_dat)
}
\arguments{
\item{prot_dat}{A DataFrame from read_maxquant_prot() with only numeric value.}
}
\value{
A DataFrame.
}
\description{
Normalize MS TMT data with Trimmed Mean of -values.
}
\details{
Input dataframe from read_maxquant_prot() function, then return a data set after normalization with TMM.
}
