% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{data_pearson_draw}
\alias{data_pearson_draw}
\title{Draw Pearson Plot.}
\usage{
data_pearson_draw(prot_dat, condition_num, repeat_num, state = "")
}
\arguments{
\item{prot_dat}{A DataFrame from with only numeric value.}

\item{condition_num}{A integer of condition number.}

\item{repeat_num}{A integer of repeat number.}

\item{state}{A title.}
}
\description{
Draw Pearson Plot.
}
\details{
Input dataframe, then return a set of Pearson plots.
}
