% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simple_analysis.R
\name{data_proteins_summary}
\alias{data_proteins_summary}
\title{Plot TMT Identified Protein Summary}
\usage{
data_proteins_summary(prot_raw, meta)
}
\arguments{
\item{prot_raw}{A DataFrame from read_maxquant_prot().}

\item{meta}{A DataFrame of meta file.}
}
\value{
A list of two ggplot objects.
}
\description{
Plot TMT identified protein summary.
}
\details{
Input dataframe from read_maxquant_prot() function, then return a list of two ggplot objects.
}
