% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{data_scale_mat}
\alias{data_scale_mat}
\title{Scale Dataframe.}
\usage{
data_scale_mat(prot)
}
\arguments{
\item{prot}{A DataFrame with only numeric value.}
}
\description{
Scale dataframe by row to mean of 0 and sd of 1.
}
\details{
Input dataframe, then return a dataframe after scaled.
}
