% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_exam.R
\docType{data}
\name{meta_exam}
\alias{meta_exam}
\title{meta file example}
\format{
A data frame with 18 rows and 6 variables:
\describe{
  \item{sample}{sample names}
  \item{condition}{experiment conditions}
  \item{repeat}{experiment repeats}
  \item{reference}{1 for reference and NA for not}
  \item{set}{experiment sets}
  \item{channel}{experiment channels the MaxQuant reporter column names}
}
}
\usage{
meta_exam
}
\description{
meta file example for a 6-plex TMT data set.
}
\keyword{datasets}
