% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_dat_specific_exam.R
\docType{data}
\name{prot_dat_specific_exam}
\alias{prot_dat_specific_exam}
\title{TMT dataset from data_transform_melt function}
\format{
A data frame with 1995 rows and 4 variables:
\describe{
  \item{id}{id of proteins}
  \item{variable}{samples}
  \item{value}{intensity}
  \item{condition}{conditions}
}
}
\usage{
prot_dat_specific_exam
}
\description{
A TMT dataset from data_transform_melt function with 5 conditions.
}
\keyword{datasets}
