% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_mean_sd_example.R
\docType{data}
\name{prot_mean_sd_example}
\alias{prot_mean_sd_example}
\title{TMT mean and sd data set}
\format{
A data frame with 133 rows and 10 variables:
\describe{
  \item{mean_0hr}{mean of 0hr condition}
  \item{sd_0hr}{sd of 0hr condition}
  \item{mean_3hr}{mean of 3hr condition}
  \item{sd_3hr}{sd of 3hr condition}
  \item{mean_9hr}{mean of 9hr condition}
  \item{sd_9hr}{sd of 9hr condition}
  \item{mean_10.5hr}{mean of 10.5hr condition}
  \item{sd_10.5hr}{sd of 10.5hr condition}
  \item{mean_12hr}{mean of 12hr condition}
  \item{sd_12hr}{sd of 12hr condition}
}
}
\usage{
prot_mean_sd_example
}
\description{
TMT mean and sd data set by conditions.
}
\keyword{datasets}
