% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_norm_example.R
\docType{data}
\name{prot_norm_example}
\alias{prot_norm_example}
\title{TMT dataset of 5 conditions and 3 repeats}
\format{
A data frame with 133 rows and 15 variables:
\describe{
  \item{0hr_re1}{reporter intensity of 0hr repeat 1}
  \item{3hr_re1}{reporter intensity of 0hr repeat 1}
  \item{9hr_re1}{reporter intensity of 0hr repeat 1}
  \item{10.5hr_re1}{reporter intensity of 0hr repeat 1}
  \item{12hr_re1}{reporter intensity of 0hr repeat 1}
  \item{0hr_re2}{reporter intensity of 0hr repeat 2}
  \item{3hr_re2}{reporter intensity of 0hr repeat 2}
  \item{9hr_re2}{reporter intensity of 0hr repeat 2}
  \item{10.5hr_re2}{reporter intensity of 0hr repeat 2}
  \item{12hr_re2}{reporter intensity of 0hr repeat 2}
  \item{0hr_re3}{reporter intensity of 0hr repeat 3}
  \item{3hr_re3}{reporter intensity of 0hr repeat 3}
  \item{9hr_re3}{reporter intensity of 0hr repeat 3}
  \item{10.5hr_re3}{reporter intensity of 0hr repeat 3}
  \item{12hr_re3}{reporter intensity of 0hr repeat 3}
}
}
\usage{
prot_norm_example
}
\description{
A dataset containing TMT dataset of 5 conditions and 3 repeats after normalization.
}
\keyword{datasets}
