% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_and_clean.R
\name{read_maxquant_prot}
\alias{read_maxquant_prot}
\title{Read, clean and annotate MS TMT data.}
\usage{
read_maxquant_prot(filename, metafile, total = FALSE)
}
\arguments{
\item{filename}{A string of processed file path.}

\item{metafile}{A string of metafile path.}

\item{total}{A boolean of considering zero row or not.}
}
\value{
A DataFrame.
}
\description{
Read, clean and annotate MS TMT data with Maxquant process data and meta file.
}
\details{
Input MaxQuang processed MS TMT data and meta file, then return a data set with annotation and TMT quantification results.
}
