% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_and_clean.R
\name{read_maxquant_prot}
\alias{read_maxquant_prot}
\title{Read, clean and annotate MS TMT data.}
\usage{
read_maxquant_prot(prot_raw, meta, total = FALSE)
}
\arguments{
\item{prot_raw}{A DataFrame of proteinGroups from MaxQuant.}

\item{meta}{A DataFrame of metafile.}

\item{total}{A boolean of considering zero row or not.}
}
\value{
A DataFrame.
}
\description{
Read, clean and annotate MS TMT data with Maxquant process data and meta file.
}
\details{
Input MaxQuang processed MS TMT data and meta file, then return a data set with annotation and TMT quantification results.
}
\examples{
meta_exam_path <- system.file("extdata", "meta.txt", package = "MSAnalysisTMTcyl")
meta_exam <- read.table(meta_exam_path,sep='\t',header=TRUE)
prot_exam_path <- system.file("extdata", "proteinGroups.txt", package = "MSAnalysisTMTcyl")
prot_exam <- read.table(prot_exam_path,sep='\t',header=TRUE)
prot_dat_example <- read_maxquant_prot(prot_exam,meta_exam)
}
